+++ anyblok/registry.py
@@ -6,7 +6,8 @@
 # This Source Code Form is subject to the terms of the Mozilla Public License,
 # v. 2.0. If a copy of the MPL was not distributed with this file,You can
 # obtain one at http://mozilla.org/MPL/2.0/.
-from os.path import join, exists
+from os.path import join
+from os import walk
 from logging import getLogger
 import nose
 
@@ -988,30 +989,29 @@ class Registry:
             funct(self.session)
 
     def run_test(registry, blok2install):
-        defaultTest = join(BlokManager.getPath(blok2install), 'tests')
-        if exists(defaultTest):
+        startpath = BlokManager.getPath(blok2install)
 
-            class ContextSuite(nose.suite.ContextSuite):
+        class ContextSuite(nose.suite.ContextSuite):
 
-                def __init__(self, *args, **kwargs):
-                    super(ContextSuite, self).__init__(*args, **kwargs)
-                    if self.context is not None:
-                        self.context.registry = registry
+            def __init__(self, *args, **kwargs):
+                super(ContextSuite, self).__init__(*args, **kwargs)
+                if self.context is not None:
+                    self.context.registry = registry
 
-            class ContextSuiteFactory(nose.suite.ContextSuiteFactory):
-                suiteClass = ContextSuite
+        class ContextSuiteFactory(nose.suite.ContextSuiteFactory):
+            suiteClass = ContextSuite
 
-            class TestLoader(nose.loader.TestLoader):
+        class TestLoader(nose.loader.TestLoader):
 
-                def __init__(self, *args, **kwargs):
-                    super(TestLoader, self).__init__(*args, **kwargs)
-                    self.suiteClass = ContextSuiteFactory(config=self.config)
+            def __init__(self, *args, **kwargs):
+                super(TestLoader, self).__init__(*args, **kwargs)
+                self.suiteClass = ContextSuiteFactory(config=self.config)
 
-            nose.run(defaultTest=[defaultTest], testLoader=TestLoader(),
-                     argv=['-v', '-s'])
-        else:
-            logger.warning("Blok %r has no %r directory" % (
-                blok2install, defaultTest))
+        for root, dirs, _ in walk(startpath):
+            if 'tests' in dirs:
+                nose.run(defaultTest=[join(root, 'tests')],
+                         testLoader=TestLoader(),
+                         argv=['-v', '-s'])
 
     def assemble_entries(self):
         for entry in RegistryManager.declared_entries: