+++ doc/CHANGES.rst
@@ -39,6 +39,8 @@ CHANGELOG
   not only at the root of the blok
 * [REF] The options **test-blok-at-install** take the **tests** directory in any 
   location in the blok not only at the root of the blok
+* [REF] The anyblok_nose console script to take the **tests** directories in any 
+  location in the blok not only at the root of the blok
 
 0.16.1 (2018-01-29)
 -------------------